# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RequestUploadUrlBody {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export type RequestUploadUrlResponseUploadParams = { [key: string]: unknown };

export interface RequestUploadUrlResponse {
  uploadURL: string;
  uploadParams: RequestUploadUrlResponseUploadParams;
  metadata?: RequestUploadUrlBody;
}

export interface ErrorEnvelope {
  error: string;
}