# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Request a presigned URL for file upload
 */





export const RequestUploadUrlBody = zod.object({
  "name": zod.string().min(1),
  "size": zod.number().min(1),
  "contentType": zod.string().min(1)
})






export const RequestUploadUrlResponse = zod.object({
  "uploadURL": zod.string(),
  "uploadParams": zod.record(zod.string(), zod.unknown()),
  "metadata": zod.object({
  "name": zod.string().min(1),
  "size": zod.number().min(1),
  "contentType": zod.string().min(1)
}).optional()
})