# lib/api-zod/src/generated/types/requestUploadUrlResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RequestUploadUrlBody } from './requestUploadUrlBody';
import type { RequestUploadUrlResponseUploadParams } from './requestUploadUrlResponseUploadParams';

export interface RequestUploadUrlResponse {
  uploadURL: string;
  uploadParams: RequestUploadUrlResponseUploadParams;
  metadata?: RequestUploadUrlBody;
}